+++ packages/crossplatform/src/helpers/platformStyleTranslation.ts
@@ -4,40 +4,11 @@ import cssToReactNative from 'css-to-react-native'
 import type React from 'react'
 import type {TextStyle, ViewStyle} from 'react-native'
 
+import {validReactNativeStyleProperties} from '../constants/validReactNativeStyleProperties'
+
 export type CrossPlatformStyle =
   | MergeWithOverride<React.CSSProperties, ViewStyle | TextStyle>
   | undefined
-
-const cssPropertiesThatCanContainRem = [
-  'width',
-  'height',
-  'font-size',
-  'line-height',
-  'top',
-  'right',
-  'bottom',
-  'left',
-  'margin',
-  'margin-top',
-  'margin-right',
-  'margin-bottom',
-  'margin-left',
-  'padding',
-  'padding-top',
-  'padding-right',
-  'padding-bottom',
-  'padding-left',
-  'border-radius',
-  'border-width',
-  'border-top-width',
-  'border-right-width',
-  'border-bottom-width',
-  'border-left-width',
-  'outline-width',
-  'letter-spacing',
-  'word-spacing',
-  'text-indent'
-]
 
 const remToPX = (cssStyles: [string, string][]): [string, string][] => {
   const remStringToPXString = (remValue: string) => {
@@ -56,12 +27,9 @@ const remToPX = (cssStyles: [string, string][]): [string, string][] => {
   }
 
   return cssStyles.map(([key, value]) => {
-    console.log('key', key)
-    console.log('value', value)
     if (
       typeof key === 'string' &&
       typeof value === 'string' &&
-      cssPropertiesThatCanContainRem.includes(key) &&
       value.includes('rem')
     ) {
       return [key, remStringToPXString(value)]
@@ -74,21 +42,96 @@ const remToPX = (cssStyles: [string, string][]): [string, string][] => {
 const numberToPX = (cssStyles: [string, string][]): [string, string][] => {
   return cssStyles.map(([key, value]) => {
     if (typeof value === 'number') {
-      return [key, `${value}px`]
+      return [key, `${value}`]
     }
 
     return [key, value]
   })
 }
 
-const webToNative = (cssStyles: [string, string][]) => {
-  return cssToReactNative(numberToPX(remToPX(cssStyles)))
+const totallyFlattenObject = (
+  obj: {[key: string]: any},
+  flattened: {[key: string]: any} = {}
+) => {
+  for (const key in obj) {
+    if (
+      typeof obj[key] === 'object' &&
+      obj[key] !== null &&
+      !Array.isArray(obj[key])
+    ) {
+      totallyFlattenObject(obj[key], flattened)
+    } else {
+      flattened[key] = obj[key]
+    }
+  }
+  return flattened
+}
+
+const webToNative = (_cssStyles: CrossPlatformStyle) => {
+  if (!_cssStyles) return {}
+  // We need to pull out the style properties that sometimes
+  // come nested depending on how the styles are defined
+  const cssStyles = totallyFlattenObject(_cssStyles)
+
+  // css-to-react-native doesn't play well with certain
+  // styles and formatting, so this is a workaround
+  const stylesToFormat = {}
+  Object.entries(cssStyles).forEach(([key, value]) => {
+    if (!isNaN(parseInt(key))) {
+      return
+    }
+    if (typeof value === 'object') {
+      return
+    }
+    if (value === undefined || key === 'mask') {
+      return
+    }
+
+    // TODO: are there any more special cases?
+    // zIndex is a special case, it needs to be a number
+    if (key !== 'zIndex' && typeof value === 'number') {
+      value = `${value}px`
+    }
+
+    Object.assign(stylesToFormat, {[key]: value})
+  })
+
+  //
+  // 1. Convert remToPX with 16px root font size
+  // 2. Convert any remaining numbers to strings since css-to-react-native doesn't do numbers well
+  // 3. Convert the css styles to react native styles
+  //
+  const formattedStyles = cssToReactNative(
+    numberToPX(remToPX(Object.entries(stylesToFormat)))
+  )
+
+  const combinedStyles = {...cssStyles, ...formattedStyles}
+
+  // RN doesn't like undefined values, also isn't a fan of some nativewind
+  // styles, so we prune them here.
+  //
+  // TODO: How do we just run this against a list of all the styles RN doesn't like?
+  const prunedCombinedStyles = Object.entries(combinedStyles).reduce(
+    (acc, [key, value]) => {
+      if (value === undefined) {
+        return acc
+      }
+
+      if (!validReactNativeStyleProperties.includes(key)) {
+        return acc
+      }
+      return {...acc, [key]: value}
+    },
+    {}
+  )
+
+  return prunedCombinedStyles
 }
 
 export const platformStyleTranslation = (cssStyle: CrossPlatformStyle) => {
   if (!cssStyle || Object.keys(cssStyle).length === 0) return {}
   if (getEnvironmentName() === 'native') {
-    return webToNative(Object.entries(cssStyle))
+    return webToNative(cssStyle)
   }
   return cssStyle
 }