+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @interunit/modal
 
+## 0.0.5
+
+### Patch Changes
+
+- - Better ergonomics for the Popover component
+  - Updated Popover docs
+  - Improved platformStyleTranslation + NativeWind dry run
+  - useOutsideClick no-op for native until support is created
+- Updated dependencies
+  - @interunit/primitives@0.0.6
+  - @interunit/config@0.0.3
+  - @interunit/a11y@0.0.6
+
 ## 0.0.4
 
 ### Patch Changes