+++ packages/primitives/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @interunit/primitives
 
+## 0.0.6
+
+### Patch Changes
+
+- - Better ergonomics for the Popover component
+  - Updated Popover docs
+  - Improved platformStyleTranslation + NativeWind dry run
+  - useOutsideClick no-op for native until support is created
+- Updated dependencies
+  - @interunit/crossplatform@0.0.2
+  - @interunit/config@0.0.3
+
 ## 0.0.5
 
 ### Patch Changes